Change check as never returns null
